torch42

+++ optimize.py
@@ -73,7 +73,7 @@ else:
 G = nx.convert_node_labels_to_integers(G)
 
 print("Converting graph to PyTorch Geometric dataset...")
-pyg_graph = pyg.utils.from_networkx(G)
+pyg_graph = pyg.utils.from_networkx(G, dtype=torch.float32)
 pyg_graph.to(device)
 
 def objective(trial):
